+++ example/lib/devices_list/devices_bloc.dart
@@ -11,28 +11,29 @@ import 'package:rxdart/rxdart.dart';
 class DevicesBloc {
   final List<BleDevice> bleDevices = <BleDevice>[];
 
-  BehaviorSubject<List<BleDevice>> _visibleDevicesController =
-      BehaviorSubject<List<BleDevice>>.seeded(<BleDevice>[]);
+  BehaviorSubject<List<BleDevice>> _visibleDevicesController = BehaviorSubject<List<BleDevice>>.seeded(<BleDevice>[]);
 
-  StreamController<BleDevice> _devicePickerController =
-      StreamController<BleDevice>();
+  StreamController<BleDevice> _devicePickerController = StreamController<BleDevice>();
 
-  StreamSubscription<ScanResult> _scanSubscription;
-  StreamSubscription _devicePickerSubscription;
+  StreamSubscription<ScanResult>? _scanSubscription;
+  StreamSubscription<BleDevice>? _devicePickerSubscription;
 
-  ValueObservable<List<BleDevice>> get visibleDevices =>
-      _visibleDevicesController.stream;
+  ValueStream<List<BleDevice>> get visibleDevices => _visibleDevicesController.stream;
 
   Sink<BleDevice> get devicePicker => _devicePickerController.sink;
 
-  DeviceRepository _deviceRepository;
-  BleManager _bleManager;
-  PermissionStatus _locationPermissionStatus = PermissionStatus.unknown;
+  final DeviceRepository _deviceRepository;
+  final BleManager _bleManager;
 
   Stream<BleDevice> get pickedDevice => _deviceRepository.pickedDevice
-      .skipWhile((bleDevice) => bleDevice == null);
+      .skipWhile((bleDevice) => bleDevice == null).cast<BleDevice>();
 
-  DevicesBloc(this._deviceRepository, this._bleManager);
+  DevicesBloc({
+    DeviceRepository? deviceRepository, 
+    BleManager? bleManager
+  }) 
+  : _deviceRepository = deviceRepository ?? DeviceRepository(),
+    _bleManager = bleManager ?? BleManager();
 
   void _handlePickedDevice(BleDevice bleDevice) {
     _deviceRepository.pickDevice(bleDevice);
@@ -40,7 +41,7 @@ class DevicesBloc {
 
   void dispose() {
     Fimber.d("cancel _devicePickerSubscription");
-    _devicePickerSubscription.cancel();
+    _devicePickerSubscription?.cancel();
     _visibleDevicesController.close();
     _devicePickerController.close();
     _scanSubscription?.cancel();
@@ -49,19 +50,12 @@ class DevicesBloc {
   void init() {
     Fimber.d("Init devices bloc");
     bleDevices.clear();
-    _bleManager
-        .createClient(
-            restoreStateIdentifier: "example-restore-state-identifier",
-            restoreStateAction: (peripherals) {
-              peripherals?.forEach((peripheral) {
-                Fimber.d("Restored peripheral: ${peripheral.name}");
-              });
-            })
-        .catchError((e) => Fimber.d("Couldn't create BLE client", ex: e))
-        .then((_) => _checkPermissions())
-        .catchError((e) => Fimber.d("Permission check error", ex: e))
-        .then((_) => _waitForBluetoothPoweredOn())
-        .then((_) => _startScan());
+
+    maybeCreateClient()
+      .then((_) => _checkPermissions())
+      .catchError((e) => Fimber.d("Permission check error", ex: e))
+      .then((_) => _waitForBluetoothPoweredOn())
+      .then((_) => _startScan());
 
     if (_visibleDevicesController.isClosed) {
       _visibleDevicesController =
@@ -73,18 +67,42 @@ class DevicesBloc {
     }
 
     Fimber.d(" listen to _devicePickerController.stream");
-    _devicePickerSubscription =
-        _devicePickerController.stream.listen(_handlePickedDevice);
+    _devicePickerSubscription = _devicePickerController.stream.listen(_handlePickedDevice);
+  }
+
+  Future<void> maybeCreateClient() async {
+    Fimber.d('Checking if client exists...');
+    final clientAlreadyExists = await _bleManager.isClientCreated();
+    Fimber.d('Client exists: $clientAlreadyExists');
+
+    if (clientAlreadyExists) {
+      Fimber.d("Client already exists");
+      return Future.value();
+    }
+
+    Fimber.d("Create client");
+
+    return _bleManager
+        .createClient(
+          restoreStateIdentifier: "example-restore-state-identifier",
+          restoreStateAction: (peripherals) {
+            peripherals.forEach((peripheral) {
+              Fimber.d("Restored peripheral: ${peripheral.name}");
+            });
+          }
+        )
+        .catchError((e) {
+          return Fimber.d("Couldn't create BLE client", ex: e);
+        });
   }
 
   Future<void> _checkPermissions() async {
     if (Platform.isAndroid) {
-      var permissionStatus = await PermissionHandler()
-          .requestPermissions([PermissionGroup.location]);
-
-      _locationPermissionStatus = permissionStatus[PermissionGroup.location];
-
-      if (_locationPermissionStatus != PermissionStatus.granted) {
+      var locGranted = await Permission.location.isGranted;
+      if (locGranted == false) {
+        locGranted = (await Permission.location.request()).isGranted;
+      }
+      if (locGranted == false) {
         return Future.error(Exception("Location permission not granted"));
       }
     }
@@ -92,39 +110,39 @@ class DevicesBloc {
 
   Future<void> _waitForBluetoothPoweredOn() async {
     Completer completer = Completer();
-    StreamSubscription<BluetoothState> subscription;
+    StreamSubscription<BluetoothState>? subscription;
     subscription = _bleManager
-        .observeBluetoothState(emitCurrentValue: true)
-        .listen((bluetoothState) async {
-      if (bluetoothState == BluetoothState.POWERED_ON &&
-          !completer.isCompleted) {
-        await subscription.cancel();
-        completer.complete();
-      }
-    });
+      .observeBluetoothState(emitCurrentValue: true)
+      .listen((bluetoothState) async {
+        if (bluetoothState == BluetoothState.POWERED_ON && !completer.isCompleted) {
+          await subscription?.cancel();
+          completer.complete();
+        }
+      });
+
     return completer.future;
   }
 
   void _startScan() {
-    Fimber.d("Ble client created");
-    _scanSubscription =
-        _bleManager.startPeripheralScan().listen((ScanResult scanResult) {
-      var bleDevice = BleDevice(scanResult);
-      if (scanResult.advertisementData.localName != null &&
-          !bleDevices.contains(bleDevice)) {
-        Fimber.d(
-            'found new device ${scanResult.advertisementData.localName} ${scanResult.peripheral.identifier}');
-        bleDevices.add(bleDevice);
-        _visibleDevicesController.add(bleDevices.sublist(0));
-      }
-    });
+    Fimber.d("Ble start scan");
+    _scanSubscription = _bleManager.startPeripheralScan()
+      .listen((scanResult) {
+        var bleDevice = BleDevice(scanResult);
+        if (!bleDevices.contains(bleDevice)) {
+          Fimber.d('found new device ${scanResult.advertisementData.localName} ${scanResult.peripheral.identifier}');
+          bleDevices.add(bleDevice);
+          _visibleDevicesController.add(bleDevices.sublist(0));
+        }
+      });
   }
 
   Future<void> refresh() async {
-    _scanSubscription.cancel();
     await _bleManager.stopPeripheralScan();
+    await _scanSubscription?.cancel();
     bleDevices.clear();
+
     _visibleDevicesController.add(bleDevices.sublist(0));
+
     await _checkPermissions()
         .then((_) => _startScan())
         .catchError((e) => Fimber.d("Couldn't refresh", ex: e));