+++ example/lib/repository/device_repository.dart
@@ -1,32 +1,30 @@
-
 import 'package:flutter_ble_lib_example/model/ble_device.dart';
 import 'package:rxdart/rxdart.dart';
 
 class MissingPickedDeviceException implements Exception {}
 
 class DeviceRepository {
+  static BleDevice? _bleDevice;
+  BehaviorSubject<BleDevice?> _deviceController;
 
-  static BleDevice _bleDevice;
-  BehaviorSubject<BleDevice> _deviceController;
+  static final DeviceRepository _deviceRepository =
+      DeviceRepository._internal();
 
-  static final DeviceRepository _deviceRepository = DeviceRepository
-      ._internal();
-
-  factory DeviceRepository() {
+   factory DeviceRepository() {
     return _deviceRepository;
   }
 
-  DeviceRepository._internal() {
-    _deviceController = BehaviorSubject<BleDevice>.seeded(_bleDevice);
-  }
+  DeviceRepository._internal() 
+  : _deviceController = BehaviorSubject<BleDevice?>.seeded(_bleDevice);
+  
 
-  void pickDevice(BleDevice bleDevice) {
+  void pickDevice(BleDevice? bleDevice) {
     _bleDevice = bleDevice;
     _deviceController.add(_bleDevice);
   }
 
-  ValueObservable<BleDevice> get pickedDevice => _deviceController.stream.shareValueSeeded(_bleDevice);
+  ValueStream<BleDevice?> get pickedDevice =>
+      _deviceController.stream.shareValueSeeded(_bleDevice);
 
   bool get hasPickedDevice => _bleDevice != null;
-
 }