+++ example/lib/test_scenarios/bluetooth_state_toggle_scenario.dart
@@ -1,7 +1,7 @@
 part of test_scenarios;
 
 class BluetoothStateTestScenario implements TestScenario {
-  StreamSubscription _radioStateSubscription;
+  StreamSubscription<BluetoothState>? _radioStateSubscription;
 
   @override
   Future<void> runTestScenario(Logger log, Logger errorLogger) async {
@@ -24,7 +24,7 @@ class BluetoothStateTestScenario implements TestScenario {
     try {
       await bleManager.enableRadio();
     } catch (e) {
-      errorLogger(e);
+      errorLogger(e.toString());
     }
 
     log("Enabled radio!");
@@ -46,7 +46,8 @@ class BluetoothStateTestScenario implements TestScenario {
 
   void _observeRadioState(BleManager bleManager, Logger log) async {
     await _radioStateSubscription?.cancel();
-    _radioStateSubscription = bleManager.observeBluetoothState().listen((newState) {
+    _radioStateSubscription =
+        bleManager.observeBluetoothState().listen((newState) {
       log("New radio state: $newState");
     }, onError: (error) {
       log("Error while observing radio state. Error: $error");