+++ lib/src/_constants.dart
@@ -1,4 +1,5 @@
 abstract class MethodName {
+  static const String isClientCreated = "isClientCreated";
   static const String createClient = "createClient";
   static const String destroyClient = "destroyClient";
 
@@ -97,7 +98,7 @@ abstract class ArgumentName {
   static const String isAutoConnect = "isAutoConnect";
   static const String requestMtu = "requestMtu";
   static const String refreshGatt = "refreshGatt";
-  static const String timeoutMillis = "timeoutMillis";
+  static const String timeoutMillis = "timeout";
   static const String emitCurrentValue = "emitCurrentValue";
 
   static const String serviceUuid = "serviceUuid";