+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+## 2.4.0
+
+* Add `BleManager#isClientCreated()` for checking whether native client exists
+
+
+## 2.3.2
+
+* Fix lack of disconnection event on iOS if connection failed to be established
+
+## 2.3.1
+
+* Fix connection timeout on iOS
+* Fix emitting current connection state on iOS
+* Update MBA to 0.1.7:
+    * Fix lack of disconnection event on iOS when connection fails to be established
+    * Fix not all errors being passed through onError callbacks on Android (thanks, @eliaslecomte)
+
+## 2.3.0
+
+* add `BleManager.createUnsafePeripheral()` to allow for connecting to known peripheral without launching scan first
+  **NOTE:** this change will not work with BLEmulator below 1.2.0
+
+## 2.2.9
+
+* Fixed issue with incorrectly typed Stream
+
+## 2.2.8
+
+* Formatted all sources according to dartfmt for consistency
+
+## 2.2.7
+
+* Minor code style fixes. Adjusted device connection state monitoring.
+
 ## 2.2.6
 
 * Fixed scan quick failures not being reported to the listener (race condition in scanning_mixin.dart)